requirements.txt
easyocr
loguru
matplotlib
natsort
numpy
pytesseract
qtawesome
scipy
webcolors
xmltodict
stdlib_list
PyQt6
baidu-aip
chardet
deep_translator
python-docx